requirements.txt
numpy==1.17.2
h5py==2.9.0
matplotlib==3.1.1
pandas==0.25.1
Pillow==5.3.0
scikit-learn==0.21.3
torch==1.2.0
tensorboard==2.0.0
tensorboardX==1.8
torchvision==0.4.0
torchnet==0.0.4
pyyaml==5.1.2
notebook==6.0.1
mkl==2019.0
six==1.12.0
nose==1.3.7
tqdm==4.36.1
pytorch-ignite==0.2.0
torchfcn==1.9.6
#bcolz #: pip - does not build wheels
#pygpu #: pip -> No matching distribution found for pygpu